scale/requirements.txt
appnope==0.1.0
astropy==1.2.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
batman-package==2.1.0
bls==0.0.1
certifi==2016.8.31
cycler==0.10.0
decorator==4.0.10
enum34==1.1.6
functools32==3.2.3.post2
gnureadline==6.3.3
ipython==3.2.1
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
matplotlib==1.5.3
nufftpy===0.1-git
numpy==1.11.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.9
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.10
pyzmq==15.4.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
stevedore==1.17.1
tornado==4.4.1
traitlets==4.3.0
virtualenv==15.0.3
virtualenv-clone==0.2.6
virtualenvwrapper==4.0
wcwidth==0.1.7
wheel==0.24.0
yolk3k==0.9
